+++ packages/appcd-config/test/fixtures/good-default-load.js
@@ -1,4 +1,0 @@
-module.exports = {
-	age: 25,
-	food: ['pizza', 'tacos']
-};