+++ packages/appcd-config/test/fixtures/good-load-foo.js
@@ -1,6 +1,0 @@
-module.exports = {
-	name: 'bar',
-	environment: {
-		name: 'foo'
-	}
-};