+++ packages/appcd-config/test/fixtures/good.js
@@ -1,5 +1,0 @@
-module.exports = {
-	name: 'foo',
-	age: 21,
-	food: ['pizza', 'tacos']
-};