+++ packages/appcd-config/test/fixtures/shebang-export.js
@@ -1,5 +1,0 @@
-#!/foo/bar
-
-module.exports = {
-	foo: 'bar'
-};